Docs: Update README and commit recent extensive refactoring and content expansion
- UI/UX major refactoring for main screens.
- Added Achievement system and enhanced Shop/Quiz features.
- Massive asset expansion (characters, backgrounds).
- Documentation updates for future plans.

+++ lib/core/utils/map_projection.dart
@@ -34,6 +34,6 @@ class MapProjection {
   }
 
   static double _clamp01(double value) {
-    return value.clamp(0.0, 1.0) as double;
+    return value.clamp(0.0, 1.0);
   }
 }
